backend/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
SQLAlchemy-Utils
flask-marshmallow
marshmallow-sqlalchemy
Flask-RESTful
requests
uuid
flask-cors